Move file shared.github to pipeline.source
