+++ installer/src/pipeline.source.ts
